feature add cart and checkout

+++ app/schemas/patch_book.py
@@ -1,9 +1,10 @@
 from pydantic import BaseModel, Field
 from typing import Optional
+from decimal import Decimal
 
 
 class PatchBook(BaseModel):
     title: Optional[str] = None
     author: Optional[str] = None
-    price: Optional[float] = Field(None, gt=0)
+    price: Optional[Decimal] = Field(None, gt=0)
     stock: Optional[int] = Field(None, ge=0)
